[ELYWEB-243] Split out the base implementation for SSO testing to allow further variants.
